Dont cast to float

+++ test_NISTfit-evaluation.py
@@ -116,7 +116,7 @@ if __name__=='__main__':
     # Allow for the number of threads to be provided at the command line as the argument to this script
     Nthreads_max = 8
     if len(sys.argv) == 2:
-        Nthreads_max = float(sys.argv[-1])
+        Nthreads_max = int(sys.argv[-1])
     
 	speedtest(evaluators.get_eval_poly, [120,12000],'speedup_polynomial.pdf')
     speedtest(evaluators.get_eval_decaying_exponential, [50,5,-1],
